Deprecate omc4py.classes module
